draft of a hurdle model implementation
